requirements-dev.txt
psycopg2-binary==2.8
pytest==5.4.1
mypy==0.740
ipython
requests
parameterized